+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # RELEASE NOTES
 
-## [undefined]
+## [v0.2.0]
 ### Added
 - output group and projects is managed by output repository
+
+### Changed
+- Move repositories creation from PersistentPreRun to each command
+- PresistentPreRun on root command only creates the configuration